+++ CHANGELOG.md
@@ -10,13 +10,23 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Enclosure Group
 - FC Network
 - FCoE Network
-- Interconnect Type
+- Interconnect
+- Interconnect type
 - Logical Enclosure
+- Logical interconnect
 - Logical Interconnect Group
 - Network Set
+- Server hardware
+- Server hardware type
 - Server Profile
 - Server Profile Template
+- Storage Pool
+- Storage System
+- Storage Volume
+- Storage Volume Attachment
+- Storage Volume Template
 - Task
+- Uplink set
 
 # [v1.1.0] (03-12-2019)
 ### Notes